Organize service item model to paylods
