+++ pkg/handlers/ghcapi/internal/payloads/model_to_payload.go
@@ -287,3 +287,81 @@ func PaymentServiceItems(paymentServiceItems *models.PaymentServiceItems) *ghcme
 	}
 	return &payload
 }
+
+// MTOServiceItemModel payload
+func MTOServiceItemModel(s *models.MTOServiceItem) *ghcmessages.MTOServiceItem {
+	if s == nil {
+		return nil
+	}
+
+	return &ghcmessages.MTOServiceItem{
+		ID:               handlers.FmtUUID(s.ID),
+		MoveTaskOrderID:  handlers.FmtUUID(s.MoveTaskOrderID),
+		MtoShipmentID:    handlers.FmtUUIDPtr(s.MTOShipmentID),
+		ReServiceID:      handlers.FmtUUID(s.ReServiceID),
+		ReServiceCode:    handlers.FmtString(string(s.ReService.Code)),
+		ReServiceName:    handlers.FmtStringPtr(&s.ReService.Name),
+		Reason:           handlers.FmtStringPtr(s.Reason),
+		PickupPostalCode: handlers.FmtStringPtr(s.PickupPostalCode),
+		Status:           ghcmessages.MTOServiceItemStatus(s.Status),
+		Dimensions:       MTOServiceItemDimensions(s.Dimensions),
+		CustomerContacts: MTOServiceItemCustomerContacts(s.CustomerContacts),
+		ETag:             etag.GenerateEtag(s.UpdatedAt),
+	}
+}
+
+// MTOServiceItemModels payload
+func MTOServiceItemModels(s models.MTOServiceItems) ghcmessages.MTOServiceItems {
+	serviceItems := ghcmessages.MTOServiceItems{}
+	for _, item := range s {
+		serviceItems = append(serviceItems, MTOServiceItemModel(&item))
+	}
+
+	return serviceItems
+}
+
+// MTOServiceItemDimension payload
+func MTOServiceItemDimension(d *models.MTOServiceItemDimension) *ghcmessages.MTOServiceItemDimension {
+	if d == nil {
+		return nil
+	}
+
+	return &ghcmessages.MTOServiceItemDimension{
+		ID:     *handlers.FmtUUID(d.ID),
+		Type:   ghcmessages.DimensionType(d.Type),
+		Length: *d.Length.Int32Ptr(),
+		Height: *d.Height.Int32Ptr(),
+		Width:  *d.Width.Int32Ptr(),
+	}
+}
+
+// MTOServiceItemDimensions payload
+func MTOServiceItemDimensions(d models.MTOServiceItemDimensions) ghcmessages.MTOServiceItemDimensions {
+	payload := make(ghcmessages.MTOServiceItemDimensions, len(d))
+	for _, item := range d {
+		payload = append(payload, MTOServiceItemDimension(&item))
+	}
+	return payload
+}
+
+// MTOServiceItemCustomerContact payload
+func MTOServiceItemCustomerContact(c *models.MTOServiceItemCustomerContact) *ghcmessages.MTOServiceItemCustomerContact {
+	if c == nil {
+		return nil
+	}
+
+	return &ghcmessages.MTOServiceItemCustomerContact{
+		Type:                       ghcmessages.CustomerContactType(c.Type),
+		TimeMilitary:               c.TimeMilitary,
+		FirstAvailableDeliveryDate: *handlers.FmtDate(c.FirstAvailableDeliveryDate),
+	}
+}
+
+// MTOServiceItemCustomerContacts payload
+func MTOServiceItemCustomerContacts(c models.MTOServiceItemCustomerContacts) ghcmessages.MTOServiceItemCustomerContacts {
+	payload := make(ghcmessages.MTOServiceItemCustomerContacts, len(c))
+	for _, item := range c {
+		payload = append(payload, MTOServiceItemCustomerContact(&item))
+	}
+	return payload
+}
